распределил логику по файлам
