+++ include/consecutive_work.h
@@ -1,0 +1,12 @@
+//
+// Created by Иван on 24.10.2020.
+//
+
+#ifndef C_HW2_TP_CONSECUTIVE_WORK_H
+#define C_HW2_TP_CONSECUTIVE_WORK_H
+
+#include <stdlib.h>
+
+void find_diff_consecutive(char* region, int* diff_count, int num_of_diff);
+
+#endif //C_HW2_TP_CONSECUTIVE_WORK_H
